requirements.txt
audioread==3.0.1
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
decorator==4.4.2
filelock==3.13.1
fsspec==2024.2.0
idna==3.6
imageio==2.33.1
imageio-ffmpeg==0.4.9
Jinja2==3.1.3
joblib==1.3.2
lazy_loader==0.3
librosa==0.10.1
llvmlite==0.42.0
MarkupSafe==2.1.5
more-itertools==10.2.0
moviepy==1.0.3
mpmath==1.3.0
msgpack==1.0.7
networkx==3.2.1
numba==0.59.0
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.19.3
nvidia-nvjitlink-cu12==12.3.101
nvidia-nvtx-cu12==12.1.105
openai-whisper @ git+https://github.com/openai/whisper.git@ba3f3cd54b0e5b8ce1ab3de13e32122d0d5f98ab
packaging==23.2
pillow==10.2.0
platformdirs==4.2.0
pooch==1.8.0
proglog==0.1.10
pycparser==2.21
regex==2023.12.25
requests==2.31.0
scikit-learn==1.4.0
scipy==1.12.0
six==1.16.0
soundfile==0.12.1
soxr==0.3.7
sympy==1.12
termcolor==2.4.0
threadpoolctl==3.2.0
tiktoken==0.6.0
torch==2.2.0
tqdm==4.66.1
triton==2.2.0
typing_extensions==4.9.0
urllib3==2.2.0
